% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{accessibility_summary}
\alias{accessibility_summary}
\title{PDF accessibility summary}
\usage{
accessibility_summary(file, profile = "ua1")
}
\arguments{
\item{file}{PDF file to check.}

\item{profile}{The validation profile to use. Default to \code{"ua1"} (recommended).}
}
\description{
Generates a simple summary of accessibility
issues for a given PDF.
}
