% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-compliant.R
\name{is_pdf_compliant}
\alias{is_pdf_compliant}
\title{Does this PDF comply with accessibility standards?}
\usage{
is_pdf_compliant(file, profile = "ua1", from_json = FALSE)
}
\arguments{
\item{file}{PDF file to check.}

\item{profile}{The validation profile to use. Default to \code{"ua1"} (recommended).}

\item{from_json}{Whether \code{file} is the JSON output or not. This is useful in
other functions to only call verapdf CLI once.}
}
\value{
A logical
}
\description{
Check whether a given PDF file is compliant to PDF/UA-1 (default).
}
\examples{
\dontrun{
is_pdf_compliant("report.pdf")
is_pdf_compliant("report.pdf", profile="ua2")
}
}
