% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-a.R
\name{is_pdfa_compliant}
\alias{is_pdfa_compliant}
\title{Is PDF/A compliant?}
\usage{
is_pdfa_compliant(file)
}
\arguments{
\item{file}{PDF file to check.}
}
\value{
A logical
}
\description{
Check whether a given PDF file is PDF/A compliant.
}
