% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verapdf.R
\name{verapdf}
\alias{verapdf}
\title{Call verapdf CLI}
\usage{
verapdf(...)
}
\arguments{
\item{...}{Additional params passed to the CLI}
}
\value{
output from the CLI
}
\description{
Utility to call verapdf command line interface.
}
