% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verapdf-cli.R
\name{verapdf}
\alias{verapdf}
\title{Call verapdf CLI}
\usage{
verapdf(file, write_to = NULL)
}
\arguments{
\item{file}{PDF file to check.}

\item{write_to}{Path to output JSON file. If \code{NULL}, not written.}
}
\value{
output from the CLI
}
\description{
Utility to call verapdf command line interface.
}
